+++ documentation/users.yml
@@ -1,0 +1,35 @@
+paths:
+  /api/users/register:
+    post:
+      summary: "Create a new user"
+      description: "Creates a new user in the system"
+      consumes:
+        - "application/json"
+      produces:
+        - "application/json"
+      parameters:
+        - in: "body"
+          name: "body"
+          description: "User object that needs to be added"
+          required: true
+          schema:
+            type: "object"
+            required:
+              - "firstName"
+              - "lastName"
+              - "email"
+              - "role"
+            properties:
+              firstName:
+                type: "string"
+              lastName:
+                type: "string"
+              email:
+                type: "string"
+              role:
+                type: "string"
+      responses:
+        201:
+          description: "User created"
+        400:
+          description: "Invalid input"
